Fixed location of issues in blender. Needed to select off objects, hit Ctrl-a, select "Location", and export.

+++ FirstCocos3D/RunningAverage.m
@@ -36,7 +36,7 @@
     
     if(self.values.count < self.size) {
         
-        NSLog(@"Not ready yet.  Size is %d", self.values.count);
+        NSLog(@"Not ready yet.  Size is %lu", (unsigned long)self.values.count);
         return 0.0;
     }
